requirements.txt
django==3.1.7
psycopg2-binary==2.8.6
djangorestframework==3.12.2
djangorestframework-xml==2.0.0
gunicorn==20.0.4
requests==2.25.1